quiz-generator/requirements.txt
langchain>=0.1.13
langgraph>=0.0.38
langchain-google-genai>=0.0.7
python-dotenv>=1.0.1
